Example of tracking a lines in the hough space
In this file, we call a function Capture_image that depending on the type of parameters pass to it track lines in the Hough space through a video. This function can also be used to only detect lines in an image using dense one-to-one Hough transform. For that it is enough to replace "feature_matching_video" by "mode_capture_photo".